fix(internal/jimm/controller.go): add default cloud region if none exist
When adding a controller JIMM will query the controller for all known clouds and then add those
clouds to its internal DB. If a cloud (e.g. MAAS or k8s) has no regions, JIMM would be unable to add
a model to that cloud as it does not store association between clouds and controllers. To solve the
issue we add a `default` region to clouds that do not have any regions and that cloud region will
only have one controller associated with it.