feat(internal/discharger/discharger.go): check controller for local offer users
When we import a model, its application offers are imported with it. For imported models JIMM
becomes the source of truth for access to application offers and is expected to discharge 3rd party
macaroons addressed to it about offer access.  If an offer has already been consumed by a local
user, those 3rd party caveats will contain local juju users and JIMM has no control over local
users. In this PR I propose the source of truth for those users remains  the Juju controller. So in
the discharge method we dial the controller and check, if the user still hass consume access to the
offer before discharging the 3rd party caveat.