ci: tweak integration test workflow
Our integration test workflow was a bit complicated because it allowed integration tests to be run manually or triggered via pull-request.
We had 2 steps that were conditional on how the test was triggered for starting either the development version of JIMM or using a tagged release.
Now we need to get an output from the step that starts JIMM, and to do that the step needs an ID which must be unique for each step.
Because we've never really used the manual run option, I've deleted that path and if we want to re-add it in the future, we can add it as a separate workflow.