chore: remove model migration test
The model migration test was failing in CI because our integration test uses the snapstore's latest JAAS snap plugin. In this PR we are making a change to the plugin which is needed for the test to pass.

While we could build the jaas plugin from source, I would consider that in a future change.